(Qt) Combine dialog files into one - makes binary smaller too
