+++ Makefile.common
@@ -489,11 +489,8 @@ ifeq ($(HAVE_QT), 1)
    OBJ += ui/drivers/ui_qt.o \
           ui/drivers/qt/ui_qt_load_core_window.o \
           ui/drivers/qt/gridview.o \
-          ui/drivers/qt/coreoptionsdialog.o \
+          ui/drivers/qt/qt_dialogs.o \
           ui/drivers/qt/filedropwidget.o \
-          ui/drivers/qt/coreinfodialog.o \
-          ui/drivers/qt/playlistentrydialog.o \
-          ui/drivers/qt/viewoptionsdialog.o \
           ui/drivers/qt/qt_playlist.o \
           ui/drivers/qt/thumbnaildownload.o \
           ui/drivers/qt/thumbnailpackdownload.o \
